Fix issues with node sorting.
* Extract to testable utility method
* Ensure synthetic nodes are sorted "last"
